move crawler agent upgrade to its own appservice
