+++ src/KamiYomu.Web/Infrastructure/AppServices/Interfaces/IDownloadAppService.cs
@@ -19,13 +19,7 @@ public interface IDownloadAppService
     /// <returns></returns>
     Task<Library> AddToCollectionAsync(AddItemCollection addItemCollection, CancellationToken cancellationToken);
 
-    /// <summary>
-    /// Recreates the manga download collection for the specified library, updating its information and ensuring that the latest chapters are available.
-    /// </summary>
-    /// <param name="library">The library to refresh.</param>
-    /// <param name="cancellationToken">Cancellation token to cancel the operation.</param>
-    /// <returns>The refreshed library.</returns>
-    Task<Library> RefreshCollectionAsync(Library library, CancellationToken cancellationToken);
+
     /// <summary>
     /// Removes a manga item from the collection based on the provided RemoveItemCollection model.
     /// </summary>
